pyproject.toml
[project]
name = "backtest-service"
version = "0.1.0"
description = "Backtest orchestrator for algorithmic trading platform"
requires-python = ">=3.13.7"
license = { text = "GNU General Public License v3.0" }
dependencies = [
    "fastapi==0.128.0",
    "uvicorn[standard]==0.40.0",
    "python-dotenv==1.2.2",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.29,<1",            # async Postgres driver
    "alembic>=1.14,<2",
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "httpx>=0.27,<1",              # async HTTP client (callbacks, external APIs)
    "python-ulid>=2,<4",          # time-sortable unique IDs for jobs
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "httpx",                       # for TestClient
    "ruff>=0.6,<1",
]


[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"
